requirements.txt
accelerate==1.13.0
addict==2.4.0
aiohttp==3.9.5
aiosignal==1.3.1
alabaster==0.7.12
anaconda-client==1.11.2
anaconda-navigator==2.4.0
anaconda-project==0.11.1
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==3.5.0
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
astroid==2.14.2
astropy==5.1
asttokens==2.0.5
async-timeout==4.0.3
atomicwrites==1.4.0
attrs==22.1.0
audioread==3.0.1
Automat==20.2.0
autopep8==1.6.0
Babel==2.11.0
backcall==0.2.0
backports.functools-lru-cache==1.6.4
backports.tempfile==1.0
backports.weakref==1.0.post1
bcrypt==3.2.0
beautifulsoup4==4.11.1
binaryornot==0.4.4
black==22.6.0
bleach==4.1.0
bokeh==2.4.3
boltons==23.0.0
Bottleneck==1.3.5
brotlipy==0.7.0
certifi==2024.6.2
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.0.4
clean-fid==0.1.35
click==8.3.1
cloudpickle==2.0.0
clyent==1.2.2
colorama==0.4.6
colorcet==3.0.1
colorlog==6.8.2
comm==0.1.2
conda==23.3.1
conda-build==3.24.0
conda-content-trust==0.1.3
conda-pack==0.6.0
conda-package-handling==2.0.2
conda_package_streaming==0.7.0
conda-repo-cli==1.0.75
conda-token==0.4.0
conda-verify==3.4.2
constantly==15.1.0
contourpy==1.0.5
cookiecutter==1.7.3
cos-python-sdk-v5==1.9.29
coscmd==1.8.6.34
crcmod==1.7
cryptography==39.0.1
cssselect==1.1.0
cycler==0.11.0
cytoolz==0.12.0
daal4py==2023.0.2
dask==2022.7.0
datasets==2.19.2
datashader==0.14.4
datashape==0.5.4
DateTime==5.5
debugpy==1.5.1
decorator==4.4.2
decord==0.6.0
deepspeed==0.14.3
defusedxml==0.7.1
diff-match-patch==20200713
diffusers==0.37.0
dill==0.3.8
distributed==2022.7.0
docstring-to-markdown==0.11
docutils==0.18.1
easydict==1.13
einops==0.8.0
entrypoints==0.4
et-xmlfile==1.1.0
executing==0.8.3
fastjsonschema==2.16.2
filelock==3.25.0
flake8==6.0.0
Flask==2.2.2
flit_core==3.6.0
flow-vis==0.1
fonttools==4.25.0
frozenlist==1.4.1
fsspec==2024.3.1
ftfy==6.3.1
future==0.18.3
gensim==4.3.0
glob2==0.7
gmpy2==2.1.2
greenlet==2.0.1
h11==0.16.0
h5py==3.7.0
HeapDict==1.0.1
hf-xet==1.3.2
hjson==3.1.0
holoviews==1.15.4
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.6.0
hvplot==0.8.2
hydra-colorlog==1.2.0
hydra-core==1.3.2
hyperlink==21.0.0
idna==3.4
imagecodecs==2021.8.26
imageio==2.26.0
imageio-ffmpeg==0.5.1
imagesize==1.4.1
imbalanced-learn==0.10.1
importlib_metadata==8.5.0
incremental==21.3.0
inflection==0.5.1
iniconfig==1.1.1
intake==0.6.7
intervaltree==3.1.0
iopath==0.1.10
ipykernel==6.19.2
ipython==8.10.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
isort==5.9.3
itemadapter==0.3.0
itemloaders==1.0.4
itsdangerous==2.0.1
jedi==0.18.1
jeepney==0.7.1
jellyfish==0.9.0
Jinja2==3.1.2
jinja2-time==0.2.0
jmespath==0.10.0
joblib==1.1.1
json5==0.9.6
jsonpatch==1.32
jsonpointer==2.1
jsonschema==4.17.3
jupyter==1.0.0
jupyter_client==7.3.4
jupyter-console==6.6.2
jupyter_core==5.2.0
jupyter-server==1.23.4
jupyterlab==3.5.3
jupyterlab-pygments==0.1.2
jupyterlab_server==2.19.0
jupyterlab-widgets==1.0.0
keyring==23.4.0
kiwisolver==1.4.4
kornia==0.7.2
kornia_rs==0.1.3
lazy_loader==0.4
lazy-object-proxy==1.6.0
libarchive-c==2.9
librosa==0.10.2.post1
lightning-utilities==0.11.9
llvmlite==0.39.1
locket==1.0.0
lpips==0.1.4
lxml==4.9.1
lz4==3.1.3
Markdown==3.4.1
markdown-it-py==4.0.0
MarkupSafe==2.1.1
matplotlib==3.7.0
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
mistune==0.8.4
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
mock==4.0.3
moviepy==1.0.3
mpmath==1.2.1
msgpack==1.0.3
multidict==6.0.5
multipledispatch==0.6.0
multiprocess==0.70.16
munkres==1.1.4
mypy-extensions==0.4.3
navigator-updater==0.3.0
nbclassic==0.5.2
nbclient==0.5.13
nbconvert==6.5.4
nbformat==5.9.2
nest-asyncio==1.5.6
networkx==2.8.4
ninja==1.11.1.1
nltk==3.7
notebook==6.5.2
notebook_shim==0.2.2
numba==0.56.4
numexpr==2.8.4
numpy==1.23.5
numpydoc==1.5.0
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-ml-py==12.555.43
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
openai-clip==1.0.1
opencv-python==4.10.0.84
openpyxl==3.0.10
packaging==22.0
pandas==1.5.3
pandocfilters==1.5.0
panel==0.14.3
param==1.12.3
parsel==1.6.0
parso==0.8.3
partd==1.2.0
pathlib==1.0.1
pathspec==0.10.3
patsy==0.5.3
peft==0.18.1
pep8==1.7.1
pesq==0.0.4
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pip==24.3.1
pkginfo==1.9.6
platformdirs==4.3.6
plotly==5.9.0
pluggy==1.0.0
ply==3.11
pooch==1.4.0
portalocker==2.10.1
poyo==0.5.0
prettytable==3.10.0
proglog==0.1.10
prometheus-client==0.14.1
prompt-toolkit==3.0.36
Protego==0.1.16
protobuf==3.20.3
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
py-cpuinfo==9.0.0
pyarrow==16.1.0
pyarrow-hotfix==0.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.8
pycodestyle==2.10.0
pycosat==0.6.4
pycparser==2.21
pycryptodome==3.20.0
pyct==0.5.0
pycurl==7.45.1
pydantic==2.7.4
pydantic_core==2.18.4
PyDispatcher==2.0.5
pydocstyle==6.3.0
pydub==0.25.1
pyerfa==2.0.0
pyflakes==3.0.1
Pygments==2.19.2
PyHamcrest==2.0.2
PyJWT==2.4.0
pylint==2.16.2
pylint-venv==2.3.0
pyls-spyder==0.4.0
pyodbc==4.0.34
pyOpenSSL==23.0.0
pyparsing==3.0.9
PyQt5-sip==12.11.0
pyrsistent==0.18.0
PySocks==1.7.1
pystoi==0.4.1
pytest==7.1.2
python-dateutil==2.8.2
python-lsp-black==1.2.1
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.7.1
python-slugify==5.0.2
python-snappy==0.6.1
python-speech-features==0.6
pytoolconfig==1.2.5
pytorch-lightning==1.9.0
pytz==2022.7
pyviz-comms==2.0.2
PyWavelets==1.4.1
pyxdg==0.27
PyYAML==6.0
pyzmq==23.2.0
QDarkStyle==3.0.2
qstylizer==0.2.2
QtAwesome==1.2.2
qtconsole==5.4.0
QtPy==2.2.0
queuelib==1.5.0
regex==2022.7.9
requests==2.32.2
requests-file==1.5.1
requests-toolbelt==0.9.1
rich==14.3.3
rope==1.7.0
Rtree==1.0.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
ruamel-yaml-conda==0.17.21
safetensors==0.4.3
SAM-2==1.0
scikit-image==0.19.3
scikit-learn==1.2.1
scikit-learn-intelex==20230228.214242
scipy==1.9.3
Scrapy==2.8.0
seaborn==0.12.2
SecretStorage==3.3.1
Send2Trash==1.8.0
sentencepiece==0.2.0
service-identity==18.1.0
setuptools==57.5.0
shellingham==1.5.4
sip==6.6.2
six==1.16.0
smart-open==5.2.1
sniffio==1.2.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.3.2.post1
soxr==0.3.7
Sphinx==5.0.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
spyder==5.4.1
spyder-kernels==2.4.1
SQLAlchemy==1.4.39
stack-data==0.2.0
statsmodels==0.13.5
supervision==0.24.0
sympy==1.13.1
tables==3.7.0
tabulate==0.8.10
TBB==0.2
tblib==1.7.0
tenacity==8.0.1
tensorboardX==2.6.2.2
terminado==0.17.1
text-unidecode==1.3
textdistance==4.2.1
threadpoolctl==2.2.0
three-merge==0.1.1
tifffile==2021.7.2
timm==1.0.10
tinycss2==1.2.1
tldextract==3.2.0
tokenizers==0.22.2
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.1
toolz==0.12.0
torch==2.6.0
torch-ema==0.3
torch_fidelity==0.4.0
torchaudio==2.6.0
torchmetrics==1.6.1
torchvision==0.21.0
tornado==6.1
tqdm==4.66.5
traitlets==5.7.1
transformers==5.3.0
triton==3.2.0
Twisted==22.2.0
typer==0.24.1
typing_extensions==4.12.2
ujson==5.4.0
Unidecode==1.2.0
urllib3==1.26.14
w3lib==1.21.0
watchdog==2.1.6
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.58.0
Werkzeug==2.2.2
whatthepatch==1.0.2
wheel==0.44.0
widgetsnbextension==3.5.2
wrapt==1.14.1
wurlitzer==3.0.2
xarray==2022.11.0
xmltodict==0.13.0
xxhash==3.4.1
yapf==0.40.2
yarl==1.9.4
zict==2.1.0
zipp==3.20.2
zope.interface==5.4.0
zstandard==0.19.0
